pyproject.toml
[tool]
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "netrix", email = "netrix89@gmail.com"},
]
dependencies = [
    "flask>=2.3.2",
    "Flask-Pydantic>=0.11.0",
    "Flask-SQLAlchemy>=3.0.3",
    "mysql-connector-python>=8.0.33",
    "flask-login>=0.6.2",
    "hashids>=1.3.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "flake8-pyproject>=1.2.3",
    "mypy>=1.3.0",
    "tox>=4.5.1",
]

[tool.pdm.scripts]
start = "flask --app l2dkp.app.main run --debug"
init_db = {call = "l2dkp.app.main:init_db"}

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
ignore = []
per-file-ignores = []
max-line-length = 100
count = true